moved authentication, collapsed response_handlers into _deserialize
