--- sdk/tables/azure-data-tables/azure/data/tables/_shared/parser.py
@@ -1,20 +1,0 @@
-# -------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License. See License.txt in the project root for
-# license information.
-# --------------------------------------------------------------------------
-
-import sys
-
-if sys.version_info < (3,):
-    def _str(value):
-        if isinstance(value, unicode):  # pylint: disable=undefined-variable
-            return value.encode('utf-8')
-
-        return str(value)
-else:
-    _str = str
-
-
-def _to_utc_datetime(value):
-    return value.strftime('%Y-%m-%dT%H:%M:%SZ')
