+++ sdk/tables/azure-data-tables/azure/data/tables/_table_service_client.py
@@ -17,7 +17,7 @@ from ._generated.models import TableProperties, TableServiceProperties, QueryOpt
 from ._models import TablePropertiesPaged, service_stats_deserialize, service_properties_deserialize
 from ._base_client import parse_connection_str, TransportWrapper
 from ._shared.models import LocationMode
-from ._shared.response_handlers import process_table_error
+from ._deserialize import process_table_error
 from ._version import VERSION
 
 from ._table_client import TableClient
