+++ sdk/tables/azure-data-tables/azure/data/tables/aio/_table_service_client_async.py
@@ -21,7 +21,7 @@ from azure.data.tables._generated.models import TableServiceProperties, TablePro
 from azure.data.tables._models import service_stats_deserialize, service_properties_deserialize
 from ._base_client_async import AsyncStorageAccountHostsMixin, AsyncTransportWrapper
 from azure.data.tables.aio._policies_async import ExponentialRetry
-from azure.data.tables._shared.response_handlers import process_table_error
+from azure.data.tables._deserialize import process_table_error
 from azure.data.tables.aio._table_client_async import TableClient
 from ._models import TablePropertiesPaged
 from .._shared._error import _validate_table_name
